+++ packages/eas-cli/src/utils/profiles.ts
@@ -7,6 +7,8 @@ import {
   SubmitProfile,
 } from '@expo/eas-json';
 
+import Log, { learnMore } from '../log';
+
 type EasProfile<T extends ProfileType> = T extends 'build'
   ? BuildProfile<Platform>
   : SubmitProfile<Platform>;
@@ -57,11 +59,15 @@ async function readProfileAsync<T extends ProfileType>({
   profileName?: string;
 }): Promise<EasProfile<T>> {
   if (type === 'build') {
-    return (await EasJsonUtils.getBuildProfileAsync(
+    const buildProfile = await EasJsonUtils.getBuildProfileAsync(
       easJsonAccessor,
       platform,
       profileName
-    )) as EasProfile<T>;
+    );
+
+    maybePrintBuildProfileDeprecationWarnings(buildProfile, profileName);
+
+    return buildProfile as EasProfile<T>;
   } else {
     return (await EasJsonUtils.getSubmitProfileAsync(
       easJsonAccessor,
@@ -69,4 +75,33 @@ async function readProfileAsync<T extends ProfileType>({
       profileName
     )) as EasProfile<T>;
   }
+}
+
+let hasPrintedDeprecationWarnings = false;
+
+function maybePrintBuildProfileDeprecationWarnings(
+  buildProfile: BuildProfile<Platform>,
+  profileName?: string
+): void {
+  if (hasPrintedDeprecationWarnings) {
+    return;
+  }
+  const deprecationWarnings = EasJsonUtils.getBuildProfileDeprecationWarnings(
+    buildProfile,
+    profileName
+  );
+  if (deprecationWarnings.length === 0) {
+    return;
+  }
+  Log.newLine();
+  Log.warn('Detected deprecated fields in eas.json:');
+  for (const warning of deprecationWarnings) {
+    const warnlog: string = warning.message.map(line => `\t${line}`).join('\n');
+    Log.warn(warnlog);
+    if (warning.docsUrl) {
+      Log.warn(`\t${learnMore(warning.docsUrl)}`);
+    }
+    Log.newLine();
+  }
+  hasPrintedDeprecationWarnings = true;
 }